% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{abilityinterface}
\alias{abilityinterface}
\title{abilityinterface}
\usage{
abilityinterface(zita_par, data, model_, method, matrix_flag, prob_matrix)
}
\arguments{
\item{zita_par}{-}

\item{data}{-}

\item{model_}{-}

\item{method}{-}

\item{matrix_flag}{-}

\item{prob_matrix}{-}
}
\value{
list
}
\description{
abilityinterface
}

