% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{autoapply}
\alias{autoapply}
\title{Autoapply}
\usage{
autoapply(clist, fun, ...)
}
\arguments{
\item{clist, }{Argument that is possibly a list}

\item{fun, }{The function}

\item{...}{additional arguments}
}
\description{
Autoapply a lapply to a function if the parameter is a list instead of a parameter.
}
\keyword{internal}

