% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapseCells}
\alias{collapseCells}
\title{Collapses Cells}
\usage{
collapseCells(O, E, mincell = 1)
}
\arguments{
\item{O}{It is a list (of length number of items) with observed frequencies
of incorrect and correct answers for each item}

\item{E}{It is a list (of length number of items) with expected frequencies
of incorrect and correct answers for each item}

\item{mincell}{is the minimum number of people expected to score k or not 
answering item j}
}
\description{
Fixes very small expected frequencies for statistical orlando
}
\keyword{internal}

