% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eapinterface}
\alias{eapinterface}
\title{eapinterface
 @param zita_par -
 @param dat -
 @param e_model -
 @param matrix_flag -
 @param prob_matrix -}
\usage{
eapinterface(zita_par, dat, e_model, matrix_flag, prob_matrix)
}
\value{
list
}
\description{
eapinterface
 @param zita_par -
 @param dat -
 @param e_model -
 @param matrix_flag -
 @param prob_matrix -
}

