% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{full.pattern.expand}
\alias{full.pattern.expand}
\title{Full expansion of patterns}
\usage{
full.pattern.expand(pts, expandcol)
}
\arguments{
\item{pts}{The pattern matrix or dataframe with frequencies.}

\item{expandcol}{The frequency column.}
}
\description{
Expansion of patterns by frequency given in a specified column.
}
\keyword{internal}

