% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg}
\alias{gg}
\title{Probability for envelopes}
\usage{
gg(a, d, cp, theta)
}
\arguments{
\item{a}{Discrimination parameter}

\item{d}{Difficulty parameter}

\item{cp}{Casual success parameter}

\item{theta}{Latent trait}
}
\description{
Calculate the probability for envelopes
}
\keyword{internal}

