% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{indexPat}
\alias{indexPat}
\title{indexPat}
\usage{
indexPat(data, pats)
}
\arguments{
\item{data}{The pattern matrix}

\item{pats}{The indexed pattern ?}
}
\description{
The index of the pattern (internal object)
}

