% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{information}
\alias{information}
\title{Test or Item information}
\usage{
information(object, range, items = NULL)
}
\arguments{
\item{object}{a matrix  (con los coeficientes estimados)}

\item{range}{a interval for which the test information should be computed.}

\item{items}{the items for which the information shoulb be computed.}
}
\value{
TotalInfo the total amount of information.

RangeInfo the amount of information in the specified interval.

RangeProp the proportion of information in the specified interval.

Range the value of range argument

items the value of items argument
}
\description{
Computes the amount of test or item information.
}
\details{
The amount of information is computed as the area under the Item or Test
Information Curve. 
The function was adapted from ltm_1.0 package.
}
\examples{
#data <- simulateTest(model = "2PL", items = 20, individuals = 800) 
#fit <- irtpp(dataset = data$test,model = "2PL")
#fit <- parameter.matrix(fit$z)
#information(fit, c(-2, 0))
#information(fit, c(0, 2), items = c(3, 5))
}
\references{
Reckase, M. (2009). Multidimensional item response theory. New York: Springer.

Baker, F. B., & Kim, S. H. (Eds.). (2004). Item response theory: Parameter estimation techniques. CRC Press.
}

