% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{package}
\name{IRTpp-package}
\alias{IRTpp-package}
\title{IRTpp : Item Response Theory Implemented in R and Cpp}
\description{
IRTpp supports the IRT and MIRT methodologies by implementing 
many IRT relevant algorithms for model exploration, calibration and fitting, IRTpp
aims to be the fastest using Cpp where it is needed, focusing in extending
IRT not only to classical applications but also in new applications which sometimes 
need highly performant implementations.
}
\details{
\tabular{ll}{
Package: \tab IRTpp\cr
Type: \tab Package\cr
Version: \tab 0.2.6\cr
Date: \tab 2016-05-14\cr
License: \tab MIT + file LICENSE \cr
}
}
\section{Getting Started}{

Get started with the IRTpp package browsing the index of this documentation
if you need help the vignettes should be helpful.


The IRTpp package allows you to use the IRT methodology for simulating, analyzing and scoring tests \cr
You can browse the package vignettes to get started.
}
\author{
SICS Research Team
}
\keyword{IRT}
\keyword{MIRT}
\keyword{Psychometry}

