% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{irtpp.model}
\alias{irtpp.model}
\title{Valid model}
\usage{
irtpp.model(model, asnumber = F)
}
\arguments{
\item{model}{Model representation in a integer, string, character, or list with a model element}

\item{asnumber}{Boolean. Set to true if you need the model as a integer (i.e. To interface with cpp)}
}
\value{
model The valid string model for this model.
}
\description{
Return a valid model string for a representation of a model.
}
\examples{
## irtpp.model("1")
## irtpp.model(2)
}

