% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{irtpp.p}
\alias{irtpp.p}
\title{Probability function for all models.}
\usage{
irtpp.p(model)
}
\arguments{
\item{model}{The model to calculate the probability to}
}
\value{
A list of probability functions that are used 
according to the required model
}
\description{
Probability function for all models.
}
\keyword{internal}

