% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{llikm}
\alias{llikm}
\title{llikm}
\usage{
llikm(params, args)
}
\arguments{
\item{params}{Parameters of the llikm interface}

\item{args}{Arguments of the llikm interface}
}
\description{
Interface for the optimized EM Step function
}
\details{
llikm
funcion que se optimiza en el paso m del EM
}

