% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{loglik.f}
\alias{loglik.f}
\title{LogLikelihood of a IRT model}
\usage{
loglik.f(test, traits, z)
}
\arguments{
\item{test}{A data frame or a matrix with the responses of the test.}

\item{traits}{A vector with each individual parameter.}

\item{z}{A list with the parameters a, b and c specified by keys.
Each key must contain a vector of the item parameters for each parameter.}
}
\value{
The value of the loglikelihood.
}
\description{
LogLikelihood of a IRT model for UIRT (Fast version)
}
\seealso{
\code{\link{loglik}}
}
\keyword{internal}

