% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mapinterface}
\alias{mapinterface}
\title{mapinterface}
\usage{
mapinterface(zita_par, dat, e_model, matrix_flag, prob_matrix)
}
\arguments{
\item{zita_par}{-}

\item{dat}{-}

\item{e_model}{-}

\item{matrix_flag}{-}

\item{prob_matrix}{-}
}
\value{
list
}
\description{
mapinterface
}

