% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{The vector to normalize}
}
\description{
Normalize a set of vectors contained in a matrix
the vectors are the rows of the matrix
a simple vector is possible
using the metric space (R^n,Sigma)
}

