% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter.list}
\alias{parameter.list}
\title{Returns a parameter list from a parameter matrix.}
\usage{
parameter.list(pars, model = "3PL", dpar = NULL, cp = NULL)
}
\arguments{
\item{pars}{The parameter matrix}

\item{model}{The model whose the matrix refers to.}

\item{dpar}{if the difficulty parameter is d and not b}

\item{cp}{if the guessing parameter is transformed to real scale}
}
\description{
Returns a parameter list from a parameter matrix.
}

