% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter.matrix}
\alias{parameter.matrix}
\title{Parameter matrix form}
\usage{
parameter.matrix(pars, model = "3PL", dims = 1, byrow = F)
}
\arguments{
\item{pars}{The parameter list}

\item{model}{The model whose the list refers to.}

\item{dims}{dimensions of the model}

\item{byrow}{to interpret the data by row or not.}
}
\value{
A matrix containing the model parameters.
}
\description{
Returns a parameter matrix from a parameter list.
}
\examples{
## create the dataset
# data <- simulateTest(model = "3PL", items = 20, individuals = 200)
## extract the parameters and transform them from one list to matrix.
# zita <-data$itempars
# parameter matrix
# zita.m <- parameter.matrix(zita)
# parameter list
# zita.l <- parameter.list(zita)

}
\seealso{
\code{\link{parameter.list}}
}

