% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern.expand}
\alias{pattern.expand}
\title{Expansion of patterns}
\usage{
pattern.expand(pts)
}
\arguments{
\item{pts}{The pattern matrix or dataframe with frequencies.}
}
\value{
A matrix of patterns expanded.
}
\description{
Expansion of patterns by frequency given in the last column.
}
\examples{
## data <- simulateTest(model = "3PL", items = 20, individuals = 200)
## pattern.expand(data$test)
}
\seealso{
\code{\link{pattern.freqs}}
}

