% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern.freqs}
\alias{pattern.freqs}
\title{pattern.freqs}
\usage{
pattern.freqs(data)
}
\arguments{
\item{data}{The pattern matrix}
}
\description{
Calculates the given frequencies for the expanded pattern matrix
}

