% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{prob.2pl}
\alias{prob.2pl}
\title{2PL fast probability function}
\usage{
prob.2pl(z, theta)
}
\arguments{
\item{z}{List containing the item parameters a and d}

\item{theta}{Vector, contains the latent traits of the individual}
}
\value{
The value of the probability for the 2PL model.
}
\description{
The probability function in the 2PL model.
}
\examples{
## Simulate the test
# data <- simulateTest(model = "2PL", items = 20, individuals = 200)
## item parameters
# zita <-data$itempars
# zita <- list(zita$a, zita$b, zita$c)
## transformation of the parameter b
# zita <- model.transform(zita, model = "3PL", src = "b", target = "d")
# zita <- list(a = zita[,1], d = zita[,1], c = zita[,1])
## Latent traits
# thetas <- data$latentTraits
## Probability model
# prob.2pl(zita, thetas)
}
\seealso{
\code{\link{probability.2pl}}, \code{\link{prob.3pl}}, 
\code{\link{probability.3pl}}, \code{\link{probability.1pl}}
}

