% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{prob.3pl}
\alias{prob.3pl}
\title{3PL fast probability function}
\usage{
prob.3pl(z, theta)
}
\arguments{
\item{z}{List containing the item parameters a, d and c.}

\item{theta}{a Vector that contains the latent traits of the individual.}
}
\description{
The probability function in the 3PL models.
}
\examples{
## Simulate the test
# data <- simulateTest(model = "3PL", items = 20, individuals = 200)
## item parameters
# zita <-data$itempars
# zita <- list(zita$a, zita$b, zita$c)
## transformation of the parameter b
# zita <- model.transform(zita, model = "3PL", src = "b", target = "d")
# zita <- list(a = zita[,1], d = zita[,1], c = zita[,1])
## Latent traits
# thetas <- data$latentTraits
## Probability model
# prob.3pl(zita, thetas)
}
\seealso{
\code{\link{probability.3pl}}, \code{\link{prob.2pl}}, 
\code{\link{probability.2pl}}, \code{\link{probability.1pl}}
}

