% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{prob}
\alias{prob}
\title{Probability multidimensional case}
\usage{
prob(theta, a, d, c)
}
\arguments{
\item{theta}{the latent trait multidimensional.}

\item{a}{Discrimination parameter in the multidimensional case.}

\item{d}{Difficulty transformed parameter d in the multidimensional case.}

\item{c}{Guessing parameter in the multidimensional case.}
}
\value{
The probability value.
}
\description{
Calculates the probability in the multidimensional case.
}
\keyword{internal}

