% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{probability.1pl}
\alias{probability.1pl}
\title{1PL probability function}
\usage{
probability.1pl(z, b = z$b, theta)
}
\arguments{
\item{z}{Optional. A list with the parameter b specified by keys.}

\item{b}{The difficulty parameter. (Optional if d is specified)}

\item{theta}{The subject's latent trait.}
}
\value{
Tue value of the probability for the model 1pl.
}
\description{
The probability function in the 1PL model.
}
\examples{
## Simulate the test
# data <- simulateTest(model = "1PL", items = 20, individuals = 200)
## item parameters
# zita <-data$itempars
## Latent traits
# thetas <- data$latentTraits
## Probability model
# probability.1pl(zita, theta = thetas)
}
\seealso{
\code{\link{probability.2pl}}, \code{\link{probability.3pl}}, 
\code{\link{prob.3pl}}, \code{\link{prob.2pl}}
}

