% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{rtnorm}
\alias{rtnorm}
\title{rtnorm}
\usage{
rtnorm(n, mean = 0, sd = 1, lower = -Inf, upper = Inf)
}
\arguments{
\item{n}{: length}

\item{mean}{: mean of the normal simulated}

\item{sd}{: Standard deviation}

\item{lower}{: lower Boundary}

\item{upper}{: upper Boundary}
}
\description{
Multidimensional simulation of tests reference
Robert, C. P. Simulation of truncated normal variables. Statistics and Computing (1995) 5, 121?125
}

