% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{s_ss}
\alias{s_ss}
\title{s_ss}
\usage{
s_ss(pr, nitems, G)
}
\arguments{
\item{pr}{matriz de probabilidad}

\item{nitems}{cantidad de items considerados en la matriz de probabilidad}

\item{G}{the group number}
}
\description{
Verosimilitudes de los scores clasicos para calcular fr. esperadas en Orlando
Calcula las verosimilitudes
}

