% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test.plot}
\alias{test.plot}
\title{Plot a test or an item.}
\usage{
test.plot(z, i = NULL)
}
\arguments{
\item{z}{Item parameter list}

\item{i}{Optional, item to plot.}
}
\value{
Void, draws a plot of the ICC or the test's ICC's
}
\description{
plots a test or an item.
}
\examples{
## Simulate the test
# data <- simulateTest(model = "3PL", items = 20, individuals = 200)
## item parameters
# zita <-data$itempars
## plot of test
# test.plot(zita)
## plot of item 5
# test.plot(zita, i = 5)
}

