% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{testmulti}
\alias{testmulti}
\title{Multidimensional test}
\usage{
testmulti(nitems, ninds, dim, model)
}
\arguments{
\item{nitems}{number of items.}

\item{ninds}{number of individuals.}

\item{dim}{latent trait dimension.}

\item{model}{"1PL", "2PL" or "3PL".}
}
\value{
the test and population parameters used in the simulation.
}
\description{
Simulates a multidimensional test.
}
\seealso{
\code{\link{simulateTest}}, \code{\link{simulateTest.file}}
\code{\link{simulateTestMD}}
}
\keyword{internal}

