% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uirtestimate}
\alias{uirtestimate}
\title{uirtestimate}
\usage{
uirtestimate(data, model_, convergenceEpsilon)
}
\arguments{
\item{data}{-}

\item{model_}{-}
}
\value{
list
}
\description{
uirtestimate
}
\keyword{internal}

