% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{x2}
\alias{x2}
\title{Calcula la estadistica x2}
\usage{
x2(model, z, patterns, G, FUN)
}
\arguments{
\item{model}{el modelo implementado.}

\item{z}{Parametros de los items, el c esta en (0,1) y recibe en el siguiente orden a,d,c OJO:d}

\item{patterns}{Patrones, frecuencias y trazos}

\item{G}{es el numero de grupos}

\item{FUN}{es la funcion con la que se calcula la probabilidad esperada en cada grupo}
}
\description{
Calcula la estadistica x2
}

