% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{abilityinterface}
\alias{abilityinterface}
\title{abilityinterface
 @param zita_par -
 @param data -
 @param model_ -
 @param method -
 @param matrix_flag -
 @param prob_matrix -}
\usage{
abilityinterface(zita_par, data, model_, method, matrix_flag, prob_matrix)
}
\value{
list
}
\description{
abilityinterface
 @param zita_par -
 @param data -
 @param model_ -
 @param method -
 @param matrix_flag -
 @param prob_matrix -
}

