% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{gg}
\alias{gg}
\title{gg}
\usage{
gg(a, d, cp, theta)
}
\arguments{
\item{a}{Slope parameter}

\item{d}{Difficulty parameter}

\item{cp}{el c en todo R}

\item{theta}{Latent trait}
}
\description{
Calcula la probabilidad para envelopes
}

