% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{llikm}
\alias{llikm}
\title{Loglik in stem M}
\usage{
llikm(params, args)
}
\arguments{
\item{params}{Parameters of the llikm interface}

\item{args}{Arguments of the llikm interface}
}
\description{
Interface for the optimized EM Step function
}
\keyword{internal}

