% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mapinterface}
\alias{mapinterface}
\title{mapinterface
 @param zita_par -
 @param dat -
 @param e_model -
 @param matrix_flag -
 @param prob_matrix -}
\usage{
mapinterface(zita_par, dat, e_model, matrix_flag, prob_matrix)
}
\value{
list
}
\description{
mapinterface
 @param zita_par -
 @param dat -
 @param e_model -
 @param matrix_flag -
 @param prob_matrix -
}

