% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize}
\alias{normalize}
\title{Normalize a vector}
\usage{
normalize(x)
}
\arguments{
\item{x}{The vector to normalize}
}
\description{
Normalize a set of vectors contained in a matrix,
the vectors are the rows of the matrix. 
A simple vector is possible
using the metric space (R^n,Sigma)
}
\examples{
## Normalize a matrix
# z <- matrix(rnorm(100, 10, 3), nrow=4)
# normalize(z)
## Normalize a vector
# y <- rnorm(20, 5, 2)
# normalize(y)
}

