% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter.list}
\alias{parameter.list}
\title{Parameter list form}
\usage{
parameter.list(pars, model = "3PL", dpar = NULL, cp = NULL)
}
\arguments{
\item{pars}{The parameter matrix}

\item{model}{The model whose the matrix refers to.}

\item{dpar}{if the difficulty parameter is d and not b.}

\item{cp}{if the guessing parameter is transformed to real scale.}
}
\value{
A list containing the model parameters.
}
\description{
Returns a parameter list from a parameter matrix.
}
\examples{
## create the dataset
# data <- simulateTest(model = "3PL", items = 20, individuals = 200)
## extract the parameters and transform them from one list to matrix.
# zita <-data$itempars
# parameter matrix
# zita.m <- parameter.matrix(zita)
# parameter list
# zita.l <- parameter.list(zita)
}
\seealso{
\code{\link{parameter.matrix}}
}

