% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter.matrix}
\alias{parameter.matrix}
\title{Returns a parameter matrix from a parameter list.}
\usage{
parameter.matrix(pars, model = "3PL", dims = 1, byrow = F)
}
\arguments{
\item{pars}{The parameter list}

\item{model}{The model whose the list refers to.}

\item{dims}{dimensions of the model}

\item{byrow}{to interpret the data by row or not.}
}
\description{
Returns a parameter matrix from a parameter list.
}

