% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern.expand}
\alias{pattern.expand}
\title{pattern.expand}
\usage{
pattern.expand(pts)
}
\arguments{
\item{pts}{The pattern matrix or dataframe with frequencies.}
}
\description{
Expansion of patterns by frequency given in the last column
}

