% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pattern.freqs}
\alias{pattern.freqs}
\title{Frequencies for the expanded pattern matrix}
\usage{
pattern.freqs(data)
}
\arguments{
\item{data}{The pattern matrix.}
}
\value{
A matrix with the frequencies of patterns expanded.
}
\description{
Calculates the given frequencies for the expanded pattern matrix
}
\examples{
## data <- simulateTest(model = "3PL", items = 20, individuals = 200)
## pattern.freqs(data$test)
}
\seealso{
\code{\link{pattern.freqs}}, \code{\link{full.pattern.expand}}
}
\keyword{internal}

