% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rtnorm}
\alias{rtnorm}
\title{Truncated normal simulation}
\usage{
rtnorm(n, mean = 0, sd = 1, lower = -Inf, upper = Inf)
}
\arguments{
\item{n}{length of the vector simulated}

\item{mean}{mean of the normal simulated}

\item{sd}{Standard deviation of the normal simulated}

\item{lower}{lower boundary of the normal simulated}

\item{upper}{upper boundary of the normal simulated}
}
\value{
the simulated value of the normal truncated
}
\description{
Truncated normal simulation
}
\examples{
# rtnorm(100, 50, 12, lower = 48)
# rtnorm(100, 50, 12, lower = 48, upper = 56)
}
\references{
Robert, C. P. Simulation of truncated normal variables. Statistics and Computing (1995) 5, 121-125
}

