% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{x2}
\alias{x2}
\title{Statistical x2.}
\usage{
x2(model, z, patterns, G, FUN)
}
\arguments{
\item{model, }{The implemented model}

\item{z, }{Parameters of the items are received in the following order, d (b transformed), c.}

\item{patterns, }{Patterns, frequencies and traits.}

\item{G, }{the number of groups.}

\item{FUN, }{It is the function with which the expected probability 
is calculated in each group.}
}
\description{
Calculates the statistical x2.
}
\keyword{internal}

